@@ -4,7 +4,7 @@ class RegistrationsControllerTest < ActionController::TestCase
 
   setup do 
     @request.env["devise.mapping"] = Devise.mappings[:user]
-    @user = FactoryGirl.create(:user)
+    @user = FactoryBot.create(:user)
   end
 
   test "should show create user page" do
@@ -30,7 +30,7 @@ class RegistrationsControllerTest < ActionController::TestCase
 
   test "should not allow to change vote location deleting the user and recreating with the same email" do
     with_blocked_change_location do
-      old_user = FactoryGirl.create(:user)
+      old_user = FactoryBot.create(:user)
       old_user.confirm
       old_user.delete
 
@@ -43,7 +43,7 @@ class RegistrationsControllerTest < ActionController::TestCase
 
   test "should not allow to change vote location deleting the user and recreating with the same vat_id" do
     with_blocked_change_location do
-      old_user = FactoryGirl.create(:user)
+      old_user = FactoryBot.create(:user)
       old_user.confirm
       old_user.delete
       
@@ -56,7 +56,7 @@ class RegistrationsControllerTest < ActionController::TestCase
 
   test "should allow to change vote location when previous user has an invalid vote_town" do
     with_blocked_change_location do
-      old_user = FactoryGirl.create(:user)
+      old_user = FactoryBot.create(:user)
       old_user.delete
       old_user.skip_before_save = true
       old_user.update_attributes vote_town: "NOTICE"
@@ -69,7 +69,7 @@ class RegistrationsControllerTest < ActionController::TestCase
 
   test "should allow to change vote location when previous user has an unverified vote_town" do
     with_blocked_change_location do
-      old_user = FactoryGirl.create(:user)
+      old_user = FactoryBot.create(:user)
       old_user.delete
       old_user.skip_before_save = true
       old_user.update_attributes vote_town: "M_01_001_4"
