@@ -1,6 +1,6 @@
 # Read about factories at https://github.com/thoughtbot/factory_girl
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :election do
     title "Hola mundo"
     agora_election_id 1
@@ -9,14 +9,14 @@ FactoryGirl.define do
     ends_at "2014-09-28 17:01:18"
     server "agora"
     
-    after(:build) { |election| election.election_locations << FactoryGirl.create(:election_location, election: election) }
+    after(:build) { |election| election.election_locations << FactoryBot.create(:election_location, election: election) }
   end
 
   trait :autonomy do
     scope 1
     after(:build) do |election| 
       election.election_locations.clear
-      election.election_locations << FactoryGirl.create(:election_location, :autonomy_location, election: election)
+      election.election_locations << FactoryBot.create(:election_location, :autonomy_location, election: election)
     end
   end
   
@@ -24,7 +24,7 @@ FactoryGirl.define do
     scope 2
     after(:build) do |election| 
       election.election_locations.clear
-      election.election_locations << FactoryGirl.create(:election_location, :province_location, election: election)
+      election.election_locations << FactoryBot.create(:election_location, :province_location, election: election)
     end
   end
 
@@ -32,7 +32,7 @@ FactoryGirl.define do
     scope 3
     after(:build) do |election|
       election.election_locations.clear
-      election.election_locations << FactoryGirl.create(:election_location, :town_location, election: election)
+      election.election_locations << FactoryBot.create(:election_location, :town_location, election: election)
     end
   end
 
@@ -40,7 +40,7 @@ FactoryGirl.define do
     scope 4
     after(:build) do |election| 
       election.election_locations.clear
-      election.election_locations << FactoryGirl.create(:election_location, :island_location, election: election)
+      election.election_locations << FactoryBot.create(:election_location, :island_location, election: election)
     end
   end
 
