@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :impulsa_project_topic do
     impulsa_project nil
     impulsa_edition_topic nil
