@@ -4,7 +4,7 @@ def create_resource_and_delete_itself klass, factory, final_count
   # Collaboration, :collaboration, 0
   scenario "a logged in user should delete itself after making a #{klass}", js: true do
     assert_equal 0, klass.all.count 
-    resource = FactoryGirl.create(factory)
+    resource = FactoryBot.create(factory)
     assert_equal 1, klass.all.count 
 
     login_as(resource.user)
@@ -22,7 +22,7 @@ feature "CanUserDeleteItselfTest" do
 
   scenario "a logged in user should delete itself" do
     assert_equal 0, User.all.count 
-    user = FactoryGirl.create(:user)
+    user = FactoryBot.create(:user)
     assert_equal 1, User.all.count 
 
     login_as(user)
@@ -39,8 +39,8 @@ feature "CanUserDeleteItselfTest" do
 
   scenario "a logged in user should delete itself after making a support on a proposal", js: true do
     assert_equal 0, Support.all.count 
-    proposal = FactoryGirl.create(:proposal)
-    user = FactoryGirl.create(:user)
+    proposal = FactoryBot.create(:proposal)
+    user = FactoryBot.create(:user)
 
     login_as(user)
     visit proposals_path
