@@ -3,8 +3,8 @@ require "test_helper"
 feature "Microcredits" do
 
   scenario "new loan - anonymous user" do
-    microcredit = FactoryGirl.create(:microcredit)
-    user = FactoryGirl.build(:user)
+    microcredit = FactoryBot.create(:microcredit)
+    user = FactoryBot.build(:user)
 
     visit microcredit_path
     page.must_have_content "Ayúdanos a financiar las campañas electorales"
@@ -34,8 +34,8 @@ feature "Microcredits" do
   end
 
   scenario "new loan - logged in user" do
-    microcredit = FactoryGirl.create(:microcredit)
-    user = FactoryGirl.create(:user)
+    microcredit = FactoryBot.create(:microcredit)
+    user = FactoryBot.create(:user)
 
     login_as(user)
     visit microcredit_path
