@@ -3,8 +3,8 @@ require "test_helper"
 feature "Proposals" do
   
   scenario "Support", js: true do
-    user = FactoryGirl.create(:user)
-    proposal = FactoryGirl.create(:proposal)
+    user = FactoryBot.create(:user)
+    proposal = FactoryBot.create(:proposal)
 
     login_as(user)
 
