@@ -3,8 +3,8 @@ require 'test_helper'
 class AdminIntegrationTest < ActionDispatch::IntegrationTest
 
   setup do
-    @user = FactoryGirl.create(:user)
-    @admin = FactoryGirl.create(:user, :admin)
+    @user = FactoryBot.create(:user)
+    @admin = FactoryBot.create(:user, :admin)
   end
 
   def login user
@@ -69,7 +69,7 @@ class AdminIntegrationTest < ActionDispatch::IntegrationTest
   #  assert_equal 2, csv.count
 
   #  # should not change count with a no_newsletter_user
-  #  FactoryGirl.create(:no_newsletter_user)
+  #  FactoryBot.create(:no_newsletter_user)
   #  get '/admin/users/download_newsletter_csv'
   #  assert_response :success
   #  assert response["Content-Type"].include? "text/csv"
@@ -77,7 +77,7 @@ class AdminIntegrationTest < ActionDispatch::IntegrationTest
   #  assert_equal 2, csv.count
 
   #  # should change count with a newsletter_user
-  #  FactoryGirl.create(:newsletter_user)
+  #  FactoryBot.create(:newsletter_user)
   #  get '/admin/users/download_newsletter_csv'
   #  assert_response :success
   #  assert response["Content-Type"].include? "text/csv"
