@@ -3,7 +3,7 @@ require 'test_helper'
 class CollaborationTest < ActiveSupport::TestCase
 
   setup do
-    @collaboration = FactoryGirl.create(:collaboration, :ccc)
+    @collaboration = FactoryBot.create(:collaboration, :ccc)
     @host = Rails.application.secrets.host
   end
 
@@ -29,13 +29,13 @@ class CollaborationTest < ActiveSupport::TestCase
   end
 
   test "should national and international scopes work" do
-    c1 = FactoryGirl.create(:collaboration, :ccc)
-    c2 = FactoryGirl.create(:collaboration, :iban)
+    c1 = FactoryBot.create(:collaboration, :ccc)
+    c2 = FactoryBot.create(:collaboration, :iban)
     c2.iban_account = "ES0690000001210123456789"
     c2.iban_bic = "ESPBESMMXXX"
     c2.save
 
-    c3 = FactoryGirl.create(:collaboration, :iban)
+    c3 = FactoryBot.create(:collaboration, :iban)
     c3.iban_account = "BE62510007547061"
     c3.iban_bic = "BEXXXXX"
     c3.save
@@ -62,13 +62,13 @@ class CollaborationTest < ActiveSupport::TestCase
   end
 
   test "should .validates_not_passport work" do
-    collaboration = FactoryGirl.build(:collaboration, :foreign_user)
+    collaboration = FactoryBot.build(:collaboration, :foreign_user)
     assert_not collaboration.valid?
     assert(collaboration.errors[:user].include? "No puedes colaborar si no dispones de DNI o NIE.")
   end
 
   test "should .validates_age_over work" do
-    user = FactoryGirl.build(:user)
+    user = FactoryBot.build(:user)
     user.update_attribute(:born_at, DateTime.now-10.years)
     @collaboration.user = user
     assert_not @collaboration.valid?
@@ -198,9 +198,9 @@ class CollaborationTest < ActiveSupport::TestCase
   end
 
   test "should .calculate_bic work" do
-    ccc = FactoryGirl.create(:collaboration, :ccc)
+    ccc = FactoryBot.create(:collaboration, :ccc)
     assert_equal "ESPBESMMXXX", ccc.calculate_bic
-    iban = FactoryGirl.create(:collaboration, :iban)
+    iban = FactoryBot.create(:collaboration, :iban)
     assert_equal "ESPBESMMXXX", iban.calculate_bic
   end
 
@@ -258,7 +258,7 @@ class CollaborationTest < ActiveSupport::TestCase
   end
 
   test "should .check_spanish_bic work" do
-    ccc = FactoryGirl.create(:collaboration, :ccc)
+    ccc = FactoryBot.create(:collaboration, :ccc)
     assert_equal "ESPBESMMXXX", ccc.calculate_bic
   end
 
@@ -298,13 +298,13 @@ class CollaborationTest < ActiveSupport::TestCase
   end
 
   test "should .payment_identifier work" do
-    credit_card = FactoryGirl.create(:collaboration, :credit_card)
+    credit_card = FactoryBot.create(:collaboration, :credit_card)
     credit_card.update_attribute(:redsys_identifier, "XXXXXX")
     assert_equal credit_card.payment_identifier, "XXXXXX"
-    iban = FactoryGirl.create(:collaboration, :iban)
+    iban = FactoryBot.create(:collaboration, :iban)
     iban.update_attribute(:payment_type, 3)
     assert_equal iban.payment_identifier, "ES0690000001210123456789/ESPBESMMXXX"
-    ccc = FactoryGirl.create(:collaboration, :ccc)
+    ccc = FactoryBot.create(:collaboration, :ccc)
     assert_equal ccc.payment_identifier, "ES0690000001210123456789/ESPBESMMXXX"
   end
 
@@ -325,7 +325,7 @@ class CollaborationTest < ActiveSupport::TestCase
     @collaboration.payment_processed! order
     assert_equal 1, @collaboration.status
 
-    credit_card = FactoryGirl.create(:collaboration, :credit_card)
+    credit_card = FactoryBot.create(:collaboration, :credit_card)
     credit_card_order = credit_card.create_order Date.today
     credit_card_order.save
     credit_card.payment_processed! credit_card_order
@@ -413,7 +413,7 @@ class CollaborationTest < ActiveSupport::TestCase
   end
 
   test "should .charge! work" do
-    collaboration = FactoryGirl.create(:collaboration, :credit_card)
+    collaboration = FactoryBot.create(:collaboration, :credit_card)
     collaboration.update_attribute(:status, 2)
     order = collaboration.create_order Date.today
     order.save
@@ -634,7 +634,7 @@ phone: '666666'"
   ##############################################
 
   test "should not save collaboration if userr is not over legal age (18 years old)" do
-    user = FactoryGirl.build(:user)
+    user = FactoryBot.build(:user)
     user.update_attribute(:born_at, DateTime.now-10.years)
     @collaboration.user = user
     assert_not @collaboration.valid?
