@@ -15,30 +15,30 @@ class ElectionTest < ActiveSupport::TestCase
   end
 
   test "should scope :active work" do
-    e1 = FactoryGirl.create(:election, starts_at: DateTime.civil(1999, 2, 2, 12, 12), ends_at: DateTime.civil(2001, 2, 2, 12, 12))
+    e1 = FactoryBot.create(:election, starts_at: DateTime.civil(1999, 2, 2, 12, 12), ends_at: DateTime.civil(2001, 2, 2, 12, 12))
     assert e1.valid?
     assert_equal(Election.active.count, 0)
-    e2 = FactoryGirl.create(:election, starts_at: DateTime.civil, ends_at: DateTime.now + 2.weeks)
+    e2 = FactoryBot.create(:election, starts_at: DateTime.civil, ends_at: DateTime.now + 2.weeks)
     assert e2.valid?
     assert_equal(Election.active.count, 1)
   end
 
   test "should .is_active? work" do
     # votacion ya cerrada
-    e1 = FactoryGirl.create(:election, starts_at: DateTime.now-30.days, ends_at: DateTime.now-7.days)
+    e1 = FactoryBot.create(:election, starts_at: DateTime.now-30.days, ends_at: DateTime.now-7.days)
     assert_not e1.is_active?
 
     # votacion activa
-    e2 = FactoryGirl.create(:election, starts_at: DateTime.now-30.days, ends_at: DateTime.now+7.days)
+    e2 = FactoryBot.create(:election, starts_at: DateTime.now-30.days, ends_at: DateTime.now+7.days)
     assert e2.is_active?
 
     # votacion del futuro, todavia no esta activada
-    e3 = FactoryGirl.create(:election, starts_at: DateTime.now+30.days, ends_at: DateTime.now+90.days)
+    e3 = FactoryBot.create(:election, starts_at: DateTime.now+30.days, ends_at: DateTime.now+90.days)
     assert_not e3.is_active?
   end
 
   test "should recently_finished? work" do
-    e = FactoryGirl.create(:election)
+    e = FactoryBot.create(:election)
     e.update_attributes(starts_at: DateTime.now-90.days, ends_at: DateTime.now+7.days)
     assert_not e.recently_finished?
     e.update_attributes(ends_at: DateTime.now-30.days)
@@ -49,13 +49,13 @@ class ElectionTest < ActiveSupport::TestCase
 
   test "should .has_valid_location_for? work" do
     # Si es una eleccion estatal todos participan
-    election = FactoryGirl.create(:election)
+    election = FactoryBot.create(:election)
 
-    user = FactoryGirl.create(:user, vote_town: "m_28_079_6")
+    user = FactoryBot.create(:user, vote_town: "m_28_079_6")
     assert election.has_valid_location_for? user
 
     # si es municipal solo los que esten en ese municipio
-    election = FactoryGirl.create(:election, :town)
+    election = FactoryBot.create(:election, :town)
     assert election.has_valid_location_for? user
 
     # si es municipal no permitir a los que no esten en ese municipio
@@ -66,12 +66,12 @@ class ElectionTest < ActiveSupport::TestCase
   test "should .user_created_at_max work" do
     # crea usuarios y eleccion
     with_versioning do
-      prev_user = FactoryGirl.create(:user, vote_town: "m_28_079_6")
+      prev_user = FactoryBot.create(:user, vote_town: "m_28_079_6")
       sleep 1
-      election = FactoryGirl.create(:election, :town)
+      election = FactoryBot.create(:election, :town)
       election.user_created_at_max = DateTime.now
       sleep 1
-      post_user = FactoryGirl.create(:user, vote_town: "m_28_079_6")
+      post_user = FactoryBot.create(:user, vote_town: "m_28_079_6")
 
       # no permite participar a usuarios creados despues de la fecha limite
       assert_not election.has_valid_user_created_at? post_user
@@ -92,46 +92,46 @@ class ElectionTest < ActiveSupport::TestCase
 
   test "should .has_valid_location_for? work other scopes" do
     # estatal
-    election = FactoryGirl.create(:election_location).election
-    user = FactoryGirl.create(:user, vote_town: "m_28_079_6")
+    election = FactoryBot.create(:election_location).election
+    user = FactoryBot.create(:user, vote_town: "m_28_079_6")
     election.update_attributes(scope: 0)
     assert election.has_valid_location_for? user
 
     # autonomia
-    election = FactoryGirl.create(:election_location, :autonomy_location).election
+    election = FactoryBot.create(:election_location, :autonomy_location).election
     election.update_attributes(scope: 1)
     assert election.has_valid_location_for? user
 
-    election = FactoryGirl.create(:election_location, :autonomy_location, location: 5).election
+    election = FactoryBot.create(:election_location, :autonomy_location, location: 5).election
     election.update_attributes(scope: 1)
     assert_not election.has_valid_location_for? user
 
     # province
-    election = FactoryGirl.create(:election_location, :province_location).election
+    election = FactoryBot.create(:election_location, :province_location).election
     election.update_attributes(scope: 2)
     assert election.has_valid_location_for? user
 
-    election = FactoryGirl.create(:election_location, :province_location, location: 29).election
+    election = FactoryBot.create(:election_location, :province_location, location: 29).election
     election.update_attributes(scope: 2)
     assert_not election.has_valid_location_for? user
 
     # town
-    election = FactoryGirl.create(:election_location, :town_location).election
+    election = FactoryBot.create(:election_location, :town_location).election
     election.update_attributes(scope: 3)
     assert election.has_valid_location_for? user
 
-    election = FactoryGirl.create(:election_location, :town_location, location: 280797).election
+    election = FactoryBot.create(:election_location, :town_location, location: 280797).election
     election.update_attributes(scope: 3)
     assert_not election.has_valid_location_for? user
 
     # island
-    election = FactoryGirl.create(:election_location, :island_location).election
+    election = FactoryBot.create(:election_location, :island_location).election
     election.update_attributes(scope: 4)
     assert_not election.has_valid_location_for? user
   end
 
   test "should .scope_name work" do
-    election = FactoryGirl.create(:election)
+    election = FactoryBot.create(:election)
     election.update_attributes(scope: 0)
     assert_equal(election.scope_name, "Estatal")
     election.update_attributes(scope: 1)
@@ -143,55 +143,55 @@ class ElectionTest < ActiveSupport::TestCase
   end
 
   test "should .scoped_agora_election_id work" do
-    election = FactoryGirl.create(:election)
-    user = FactoryGirl.create(:user)
+    election = FactoryBot.create(:election)
+    user = FactoryBot.create(:user)
     assert_equal(1000, election.scoped_agora_election_id(user))
 
-    election = FactoryGirl.create(:election, :autonomy)
+    election = FactoryBot.create(:election, :autonomy)
     assert_equal(1111, election.scoped_agora_election_id(user))
 
-    election = FactoryGirl.create(:election, :province)
+    election = FactoryBot.create(:election, :province)
     assert_equal(1280, election.scoped_agora_election_id(user))
 
-    election = FactoryGirl.create(:election, :town)
+    election = FactoryBot.create(:election, :town)
     assert_equal(12807960, election.scoped_agora_election_id(user))
 
-    user = FactoryGirl.create(:user, :island)
-    election = FactoryGirl.create(:election, :island_election)
+    user = FactoryBot.create(:user, :island)
+    election = FactoryBot.create(:election, :island_election)
     assert_equal(1730, election.scoped_agora_election_id(user))
 
-    user = FactoryGirl.create(:user, :foreign_address)
-    election = FactoryGirl.create(:election, :foreign_election)
+    user = FactoryBot.create(:user, :foreign_address)
+    election = FactoryBot.create(:election, :foreign_election)
     assert_equal(1000, election.scoped_agora_election_id(user))
   end
 
   test "should full_title_for work" do
-    user = FactoryGirl.create(:user)
-    user2 = FactoryGirl.create(:user, town: "m_01_001_4")
+    user = FactoryBot.create(:user)
+    user2 = FactoryBot.create(:user, town: "m_01_001_4")
 
-    election = FactoryGirl.create(:election)
+    election = FactoryBot.create(:election)
     assert_equal("Hola mundo", election.full_title_for(user))
     assert_equal("Hola mundo", election.full_title_for(user2))
 
-    election = FactoryGirl.create(:election, :autonomy)
+    election = FactoryBot.create(:election, :autonomy)
     assert_equal("Hola mundo en Comunidad de Madrid", election.full_title_for(user))
     assert_equal("Hola mundo (no hay votación en País Vasco/Euskadi)", election.full_title_for(user2))
 
-    election = FactoryGirl.create(:election, :town)
+    election = FactoryBot.create(:election, :town)
     assert_equal("Hola mundo en Madrid", election.full_title_for(user))
     assert_equal("Hola mundo (no hay votación en Alegría-Dulantzi)", election.full_title_for(user2))
 
   end
 
   test "should locations work" do
-    election = FactoryGirl.create(:election, :town)
-    election.election_locations << FactoryGirl.create(:election_location, election: election, location: 280797, agora_version: 1)
-    election.election_locations << FactoryGirl.create(:election_location, election: election, location: 280798, agora_version: 0)
+    election = FactoryBot.create(:election, :town)
+    election.election_locations << FactoryBot.create(:election_location, election: election, location: 280797, agora_version: 1)
+    election.election_locations << FactoryBot.create(:election_location, election: election, location: 280798, agora_version: 0)
     assert_equal( "280796,0\n280797,1\n280798,0", election.locations )
   end
 
   test "should locations= work" do
-    election = FactoryGirl.create(:election)
+    election = FactoryBot.create(:election)
     election.election_locations.clear
     election.locations = "280796,0\n280797,0\n280798,0"
     election.save
@@ -212,13 +212,13 @@ class ElectionTest < ActiveSupport::TestCase
   end
 
   test "should server_shared_key work" do
-    election = FactoryGirl.create(:election)
+    election = FactoryBot.create(:election)
     assert_equal( "changeme", election.server_shared_key )
   end
 
   test "should server_url work" do
-    election = FactoryGirl.create(:election)
-    beta_election = FactoryGirl.create(:election, :beta_server)
+    election = FactoryBot.create(:election)
+    beta_election = FactoryBot.create(:election, :beta_server)
     assert_equal( "https://example.com/", election.server_url )
     assert_equal( "https://beta.example.com/", beta_election.server_url )
   end
