@@ -9,7 +9,7 @@ end
 class OrderTest < ActiveSupport::TestCase
 
   setup do
-    @collaboration = FactoryGirl.create(:collaboration, :ccc)
+    @collaboration = FactoryBot.create(:collaboration, :ccc)
     @order = @collaboration.create_order Date.today, true
     @host = Rails.application.secrets.host
   end
@@ -205,7 +205,7 @@ class OrderTest < ActiveSupport::TestCase
   end
 
   test "should .mark_bank_orders_as_charged! work" do
-    collaboration1 = FactoryGirl.create(:collaboration, :credit_card)
+    collaboration1 = FactoryBot.create(:collaboration, :credit_card)
     order1 = collaboration1.create_order Date.today, true
     order1.save
     Order.mark_bank_orders_as_charged! Date.today
@@ -214,7 +214,7 @@ class OrderTest < ActiveSupport::TestCase
     assert_equal(0, order1.status)
 
     # shouldnt mark as charged iban type collaboration order
-    collaboration2 = FactoryGirl.create(:collaboration, :iban)
+    collaboration2 = FactoryBot.create(:collaboration, :iban)
     order2 = collaboration2.create_order Date.today, true
     order2.save
     Order.mark_bank_orders_as_charged! Date.today+1.hour
@@ -229,7 +229,7 @@ class OrderTest < ActiveSupport::TestCase
     # shouldnt mark as charged ccc type collaboration order
     assert_equal(0, @order.status)
 
-    collaboration2 = FactoryGirl.create(:collaboration, :iban)
+    collaboration2 = FactoryBot.create(:collaboration, :iban)
     order2 = collaboration2.create_order Date.today, true
     order2.save
 
